Move `slice::Windows` to a new module
